# workstreams/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sovereignty Lens API
  version: 1.0.0
  summary: Audience-participation dependency graph API for the Sovereignty Lens demo.
  description: |
    HTTP API of the Sovereignty Lens backend, a standalone Java 21 / Spring Boot 3
    service backed by plain PostgreSQL.

    **All data served by this API is simulated, unverified demo data.** Organizations
    and dependencies are fictional or audience-submitted during a live presentation.
    Nothing returned here is a factual claim about any real government body, company,
    or supplier relationship, and it must never be presented as one.

    ## Contract

    Every request and response body conforms to **contract version `1`** of
    `contracts/data-contract.md`, and every body carries `contractVersion: 1`.
    Transport details for the live event stream are recorded in
    `contracts/transport-amendment.md`; the event payloads themselves are unchanged
    from contract version `1`.

    Conventions:

    - All IDs are lowercase UUID strings.
    - All timestamps are UTC RFC 3339 / ISO 8601 strings.
    - JSON is camelCase; the database is snake_case.
    - Request and event schemas are strict: unknown fields are rejected.
    - An edge always means `source organization depends on target organization`.
      Reversing that direction is a breaking contract change.

    ## Fixed error-code to HTTP-status mapping

    Every failure returns an `ApiErrorResponse`. The mapping is fixed by the data
...

...

...
servers:
...
paths:
...
  /api/sessions/{slug}/dependencies:
    post:
      tags: [Public]
      operationId: createDependency
      deprecated: true
      summary: Submit one audience dependency
      description: |
        Accepts one audience contribution: the selected source organization gains one
        dependency on a named target company.

        `anonymousClientId` is a UUID generated and persisted by the browser. The
        backend hashes it with a server-only secret before storage; it is never
        returned and never appears in an event.

        The whole submission is one database transaction. On commit the backend
        returns the canonical node and edge and publishes exactly one
        `dependency.created` event carrying the same `eventId`, `node.id` and
        `edge.id`. A rollback persists neither the dependency nor the event.

        Enforced invariants: the session must be open, the source must belong to the
        session and be active, the browser may contribute at most once per session
        round, the round holds at most 150 audience dependencies, source and target
        may not be equal, the target organization type may not be `government`, and
        one source/target edge may be active at most once per round.
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionRequest'
      responses:
        '201':
          description: The dependency was committed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionResult'
        '400':
          description: '`VALIDATION_ERROR` — invalid input, self-dependency, or a `government` target.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                validationError:
                  summary: Company name too short
                  value:
                    contractVersion: 1
                    error:
                      code: VALIDATION_ERROR
                      message: Company name must contain 2 to 60 characters.
                      retryable: false
                      field: target.name
        '404':
          description: '`SESSION_NOT_FOUND` or `SOURCE_NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                sourceNotFound:
                  summary: Source organization is not part of this session
                  value:
                    contractVersion: 1
                    error:
                      code: SOURCE_NOT_FOUND
                      message: The selected organization is not part of this session.
                      retryable: false
                      field: sourceOrganizationId
        '409':
          description: '`DUPLICATE_DEPENDENCY` or `ALREADY_CONTRIBUTED`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                alreadyContributed:
                  summary: This browser already contributed in the current round
                  value:
                    contractVersion: 1
                    error:
                      code: ALREADY_CONTRIBUTED
                      message: This device already added a dependency in the current round.
                      retryable: false
                duplicateDependency:
                  summary: The same edge is already active in this round
                  value:
                    contractVersion: 1
                    error:
                      code: DUPLICATE_DEPENDENCY
                      message: That dependency has already been added in this round.
                      retryable: false
        '423':
          description: '`SESSION_PAUSED` — the presenter paused contributions.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                sessionPaused:
                  summary: Session paused
                  value:
                    contractVersion: 1
                    error:
                      code: SESSION_PAUSED
                      message: The presenter paused contributions.
                      retryable: false
        '429':
          description: '`ROUND_CAPACITY_REACHED` — the current round already holds 150 audience dependencies.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                roundFull:
                  summary: Round at capacity
                  value:
                    contractVersion: 1
                    error:
                      code: ROUND_CAPACITY_REACHED
                      message: This round is full. Watch the presentation for the next one.
                      retryable: false
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/admin/login:
    post:
      tags: [Admin]
      operationId: adminLogin
      summary: Presenter login with the shared password
      description: |
        Validates the submitted password against the configured presenter password
        using a constant-time comparison and, on success, sets the signed
        `sl_admin` session cookie.

        Failures return a generic `UNAUTHORIZED` error and never reveal which check
        failed.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Authenticated. The signed admin session cookie is set.
          headers:
            Set-Cookie:
              description: |
                Signed presenter session cookie. Always `HttpOnly`, `SameSite=Lax`
                and `Path=/`; `Secure` is added whenever the service is served over
                HTTPS.
              schema:
                type: string
                example: sl_admin=eyJzaWQiOiJkZW1vIn0.c2lnbmF0dXJl; Path=/; HttpOnly; SameSite=Lax; Max-Age=28800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResult'
        '400':
          description: '`VALIDATION_ERROR` — missing or blank password.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/session:
    get:
      tags: [Admin]
      operationId: getAdminSession
      summary: Verify the presenter session
      description: Returns the current demo session when the signed admin cookie is valid.
      responses:
        '200':
          description: The presenter is authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/sessions/{slug}/actions:
    post:
      tags: [Admin]
      operationId: performAdminAction
      summary: Run a presenter recovery action
      description: |
        Applies one presenter action and returns the resulting authoritative session
        state.

        - `pause` — set session status to `paused`; contributions return `423`.
        - `resume` — set session status to `open`.
        - `undo` — hide the newest active audience dependency of the current round.
        - `reset` — increment `currentRound` and reopen the session. Reset never
          deletes data; seed dependencies stay visible in every round.

        Each action emits one `graph.invalidated` event on the session stream with
        the matching reason. A reset event carries the **new** current round.
      parameters:
        - $ref: '#/components/parameters/SessionSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminActionRequest'
      responses:
        '200':
          description: The action was applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminActionResult'
        '400':
          description: '`VALIDATION_ERROR` — unknown action type, or nothing to undo.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/SessionNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/admin/sessions/{slug}/dependencies:
    get:
      tags: [Admin]
      operationId: listAdminDependencies
      summary: List current-round audience dependencies for review
      description: |
        Returns all current-round, non-seed dependencies, **including hidden
        entries**, ordered newest first, each expanded with its source and target
        node so the presenter can moderate without a second lookup.
      parameters:
        - $ref: '#/components/parameters/SessionSlug'
      responses:
        '200':
          description: The moderation list for the current round.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminDependencyList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/SessionNotFound'
  /api/admin/dependencies/{id}:
    patch:
      tags: [Admin]
      operationId: updateDependencyStatus
      summary: Hide or restore one audience dependency
      description: |
        Sets the visibility of one dependency. Only a current-round, non-seed
        dependency can be modified; seed edges and edges from earlier rounds return
        `404`.

        Hiding emits `graph.invalidated` with reason `hide`, restoring with reason
        `restore`. Hidden edges disappear from public snapshots along with any nodes
        made unreachable solely by them.
      parameters:
        - name: id
          in: path
          required: true
          description: Dependency (edge) ID.
          schema:
            type: string
            format: uuid
          example: 7f8e9d0c-1b2a-4c3d-8e5f-6a7b8c9d0e1f
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DependencyStatusRequest'
      responses:
        '200':
          description: The new canonical edge.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencyStatusResult'
        '400':
          description: '`VALIDATION_ERROR` — unknown status value.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: '`NOT_FOUND` — no such current-round, non-seed dependency.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
              examples:
                notFound:
                  summary: Not a current-round audience dependency
                  value:
                    contractVersion: 1
                    error:
                      code: NOT_FOUND
                      message: No editable dependency with that ID in the current round.
                      retryable: false
security:
  - adminSession: []
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: sl_admin
      description: |
        Signed, `HttpOnly`, `SameSite=Lax` presenter session cookie issued by
        `POST /api/admin/login`. Required by every `/api/admin/**` endpoint except
        login and logout. A missing, malformed, expired, or badly signed cookie
        yields `401` with error code `UNAUTHORIZED`.
  parameters:
    SessionSlug:
      name: slug
      in: path
      required: true
      description: URL-safe session slug. The demo session is `demo`.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
      example: demo
  responses:
    Unauthorized:
      description: '`UNAUTHORIZED` — missing or invalid presenter session.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            unauthorized:
              summary: No valid admin cookie
              value:
                contractVersion: 1
                error:
                  code: UNAUTHORIZED
                  message: Presenter authentication required.
                  retryable: false
    SessionNotFound:
      description: '`SESSION_NOT_FOUND` — no session with that slug.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            sessionNotFound:
              summary: Unknown slug
              value:
                contractVersion: 1
                error:
                  code: SESSION_NOT_FOUND
                  message: No session with that slug.
                  retryable: false
                  field: slug
    InternalError:
      description: '`INTERNAL_ERROR` — unexpected server failure. Retryable when transient.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            internalError:
              summary: Transient failure
              value:
                contractVersion: 1
                error:
                  code: INTERNAL_ERROR
                  message: Temporary server problem. Please try again.
                  retryable: true
  schemas:
    ContractVersion:
      type: integer
      enum: [1]
      description: Wire contract version. Always `1` for this specification.
      example: 1
...
    ApiErrorResponse:
      type: object
      description: Uniform error envelope used by every endpoint.
      additionalProperties: false
      required: [contractVersion, error]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              $ref: '#/components/schemas/ApiErrorCode'
            message:
              type: string
              description: Human-readable message safe to show to an audience member.
              example: Company name must contain 2 to 60 characters.
            retryable:
              type: boolean
              description: |
                True only for transient `INTERNAL_ERROR` failures. Validation,
                duplicate, paused and capacity errors are never retryable.
              example: false
            field:
              type: string
              description: |
                Omitted unless the failure is attributable to exactly one request
                field. Uses dotted JSON paths.
              example: target.name
...
    ContributionRequest:
      type: object
      description: One audience submission. Unknown fields are rejected.
      additionalProperties: false
      required: [contractVersion, anonymousClientId, sourceOrganizationId, target]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        anonymousClientId:
          type: string
          format: uuid
          description: |
            UUID generated and persisted by the browser in local storage. Hashed with
            a server secret before storage; never returned and never broadcast.
          example: b1c9f0d2-7a44-4e6b-9c31-5d0e8f2a6b73
        sourceOrganizationId:
          type: string
          format: uuid
          description: An existing active organization of this session that gains the dependency.
          example: 3c4d5e6f-7081-4c9d-8eaf-2a3b4c5d6e7f
        target:
          type: object
          description: The company the source organization depends on.
          additionalProperties: false
          required: [name, organizationType, jurisdiction]
          properties:
            name:
              type: string
              minLength: 2
              maxLength: 60
              description: |
                Trimmed to 2 to 60 Unicode characters. Normalized for
                case-insensitive uniqueness inside the session; an existing match is
                reused rather than duplicated.
              example: Northwind Cloud Services
            organizationType:
              $ref: '#/components/schemas/ContributionTargetType'
            jurisdiction:
              $ref: '#/components/schemas/Jurisdiction'
      example:
        contractVersion: 1
        anonymousClientId: b1c9f0d2-7a44-4e6b-9c31-5d0e8f2a6b73
        sourceOrganizationId: 3c4d5e6f-7081-4c9d-8eaf-2a3b4c5d6e7f
        target:
          name: Northwind Cloud Services
          organizationType: cloud
          jurisdiction: united_states
    ContributionResult:
      type: object
      description: |
        Canonical persisted result. `eventId`, `node.id` and `edge.id` exactly match
        the corresponding `dependency.created` event, so a client that receives both
        applies them idempotently.
      required: [contractVersion, eventId, round, node, edge]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        eventId:
          type: string
          format: uuid
          example: 9b7d4f21-3c8e-4d6a-b5f0-7e2c1a8d9034
        round:
          type: integer
          minimum: 1
          example: 3
        node:
          $ref: '#/components/schemas/GraphNode'
        edge:
          $ref: '#/components/schemas/GraphEdge'
      example:
        contractVersion: 1
        eventId: 9b7d4f21-3c8e-4d6a-b5f0-7e2c1a8d9034
        round: 3
        node:
          id: 8a9b0c1d-2e3f-4a5b-8c7d-6e5f4a3b2c1d
          name: Northwind Cloud Services
          organizationType: cloud
          jurisdiction: united_states
          isSeed: false
        edge:
          id: 7f8e9d0c-1b2a-4c3d-8e5f-6a7b8c9d0e1f
          sourceOrganizationId: 3c4d5e6f-7081-4c9d-8eaf-2a3b4c5d6e7f
          targetOrganizationId: 8a9b0c1d-2e3f-4a5b-8c7d-6e5f4a3b2c1d
          isSeed: false
          status: active
          createdAt: '2026-08-27T10:15:30Z'
...
    AdminLoginRequest:
      type: object
      description: Shared-password presenter login. Unknown fields are rejected.
      additionalProperties: false
      required: [contractVersion, password]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        password:
          type: string
          minLength: 1
          format: password
          description: The shared presenter password. Compared in constant time; never logged.
          example: correct-horse-battery-staple
      example:
        contractVersion: 1
        password: correct-horse-battery-staple
    AdminLoginResult:
      type: object
      description: Successful login. The session itself lives in the signed `sl_admin` cookie.
      required: [contractVersion, authenticated]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        authenticated:
          type: boolean
          enum: [true]
          example: true
      example:
        contractVersion: 1
        authenticated: true
    AdminSessionResult:
      type: object
      additionalProperties: false
      required: [contractVersion, authenticated, session]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        authenticated:
          type: boolean
          enum: [true]
        session:
          $ref: '#/components/schemas/SessionSummary'
...
    AdminActionRequest:
      type: object
      description: One presenter action, wrapped in a discriminated `action` object.
      additionalProperties: false
      required: [contractVersion, action]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        action:
          type: object
          additionalProperties: false
          required: [type]
          properties:
            type:
              $ref: '#/components/schemas/AdminActionType'
      example:
        contractVersion: 1
        action:
          type: pause
    AdminActionResult:
      type: object
      description: |
        Result of a presenter action, including the resulting authoritative session
        state. `eventId` matches the emitted `graph.invalidated` event.
      required: [contractVersion, eventId, session]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        eventId:
          type: string
          format: uuid
          example: c4e2a180-5b93-4f07-9a6d-1d8b3f5e7c20
        session:
          $ref: '#/components/schemas/SessionSummary'
      example:
        contractVersion: 1
        eventId: c4e2a180-5b93-4f07-9a6d-1d8b3f5e7c20
        session:
          id: 6f1b3c2a-9d54-4a7e-8b21-2f0c5d7e9a10
          slug: demo
          title: European Digital Services Agency dependency demo
          status: paused
          currentRound: 3
          rootOrganizationId: 1a2b3c4d-5e6f-4a7b-8c9d-0e1f2a3b4c5d
    DependencyStatusRequest:
      type: object
      description: Hide or restore one current-round audience dependency.
      additionalProperties: false
      required: [contractVersion, status]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        status:
          $ref: '#/components/schemas/DependencyStatus'
      example:
        contractVersion: 1
        status: hidden
    DependencyStatusResult:
      type: object
      description: |
        Result of a hide or restore action. `eventId` matches the emitted
        `graph.invalidated` event.
      required: [contractVersion, eventId, edge]
      properties:
        contractVersion:
          $ref: '#/components/schemas/ContractVersion'
        eventId:
          type: string
          format: uuid
          example: d5f3b291-6ca4-4018-8b7e-2e9c4a6f8d31
        edge:
          $ref: '#/components/schemas/GraphEdge'
      example:
        contractVersion: 1
        eventId: d5f3b291-6ca4-4018-8b7e-2e9c4a6f8d31
        edge:
          id: 7f8e9d0c-1b2a-4c3d-8e5f-6a7b8c9d0e1f
          sourceOrganizationId: 3c4d5e6f-7081-4c9d-8eaf-2a3b4c5d6e7f
          targetOrganizationId: 8a9b0c1d-2e3f-4a5b-8c7d-6e5f4a3b2c1d
          isSeed: false
          status: hidden
          createdAt: '2026-08-27T10:15:30Z'
...